refactor: move iteration metadatas components to own dir
